+++ packages/react-server/CHANGELOG.md
@@ -7,6 +7,8 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## [Unreleased] -->
 
+- Added cookie middleware
+
 ## [0.3.1] - 2019-08-29
 
 ### Fixed